Create Analysis of Call Volume
This project focuses on applying exploratory data analysis of the Call Volume dataset to find insights and address important business questions in the call center industry. The dataset that spans 22 days comprises a wide range of information, including the agent’s name and ID, the queue time (how long a customer had to wait before connecting with an agent), the time of the call, the duration of the call, and the call status (whether it was abandoned, answered, or transferred).